fix: Init DDEShellSurfaceInterface parent QObject
DDEShellSurfaceInterface will keep the object in s_shellSurfaces, and it
will be remove when QObject::destroyed is signaled. Because the QObject
is not inited in DDEShellSurfaceInterface constructor, the signal will
not emitted, then the object in s_shellSurfaces can note be removed.

Log: Fix object in s_shellSurfaces cant not be removed
Task: https://pms.uniontech.com/bug-view-140013.html

Signed-off-by: Chaojiang Luo <luochaojiang@uniontech.com>
Change-Id: I017b0894c5dbd7a691cfd47626ed507094de3674